fix: 상품 리스트 페이지 스켈레톤과 로딩의 분리
