+++ src/app/(shop)/products/[mainCategory]/_components/ProductListInner.tsx
@@ -3,8 +3,8 @@
 import { useState } from 'react';
 import ProductsCard from '@/app/components/ProductsCard';
 import ProductsCardList from '@/app/components/ProductsCardList';
-import ProductListSkeleton from './ProductListSkeleton';
 import { Products } from '@/app/lib/products';
+import Skeleton from '../skeleton';
 
 type Props = {
   products: Products[];
@@ -24,10 +24,8 @@ export default function ProductListInner({ products, category, sort, baseUrl }: 
 
   return (
     <div className="relative">
-      {/* 데이터가 불러와지기 전까지 스켈레톤을 구현했습니다 */}
-      {!isAllLoaded && <ProductListSkeleton count={products.length || 12} />}
+      {!isAllLoaded && <Skeleton count={products.length || 12} />}
 
-      {/*상품 정보를 넣기위한 children을 받고 있습니다 */}
       <ProductsCardList
         className={`transition-opacity duration-150 ${isAllLoaded ? 'opacity-100' : 'opacity-0 absolute inset-0 pointer-events-none'}`}
       >
