+++ src/app/(shop)/products/[mainCategory]/page.tsx
@@ -4,8 +4,8 @@ import BreadCrumble from './_components/BreadCrumble';
 import Sort from './_components/Sort';
 import FilterCategory from './_components/filterCategory';
 import { Suspense } from 'react';
-import ProductListSkeleton from './_components/ProductListSkeleton';
 import ProductListFetcher from './_components/ProductListFetcher';
+import Skeleton from './skeleton';
 
 type Product = {
   params: Promise<{
@@ -40,9 +40,9 @@ export default async function ProductListPage({ params, searchParams }: Product)
         <FilterCategory mainCategory={mainCategory} category={category} sort={sort} page={page} />
         <Sort mainCategory={mainCategory} category={category} />
       </div>
-      <main>
+      <main id="main-content">
         {/* 페이지 목록 영역 입니다 Suspense로 감싸고 안에 상품에 데이터를 전달 할 수 있도록 컴포넌트를 불러와 작성해줍니다 */}
-        <Suspense fallback={<ProductListSkeleton />}>
+        <Suspense fallback={<Skeleton />}>
           {/*
            * baseurl: 초기 경로
            * page: 현재 페이지
